daemon: fix network hashrate display

+++ xelis_daemon/src/main.rs
@@ -136,7 +136,7 @@ async fn run_prompt<S: Storage>(prompt: ShareablePrompt<Arc<Blockchain<S>>>, blo
             mempool.size()
         };
 
-        let network_hashrate = (blockchain.get_difficulty() / BLOCK_TIME_MILLIS) as f64 / MILLIS_PER_SECOND as f64;
+        let network_hashrate = (blockchain.get_difficulty() / (BLOCK_TIME_MILLIS / MILLIS_PER_SECOND)) as f64;
 
         Ok(
             build_prompt_message(
@@ -376,7 +376,7 @@ async fn status<S: Storage>(manager: &CommandManager<Arc<Blockchain<S>>>, _: Arg
     manager.message(format!("Stable Height: {}", stableheight));
     manager.message(format!("Topo Height: {}", topoheight));
     manager.message(format!("Difficulty: {}", format_difficulty(difficulty)));
-    manager.message(format!("Network Hashrate: {}", format_hashrate((difficulty / BLOCK_TIME_MILLIS) as f64 / MILLIS_PER_SECOND as f64)));
+    manager.message(format!("Network Hashrate: {}", format_hashrate((difficulty / (BLOCK_TIME_MILLIS / MILLIS_PER_SECOND)) as f64)));
     manager.message(format!("Top block hash: {}", top_block_hash));
     manager.message(format!("Average Block Time: {:.2}s", avg_block_time as f64 / MILLIS_PER_SECOND as f64));
     manager.message(format!("Target Block Time: {:.2}s", BLOCK_TIME_MILLIS as f64 / MILLIS_PER_SECOND as f64));
